Address comments and Fix CI
