+++ airflow-core/tests/unit/cli/commands/test_plugins_command.py
@@ -86,8 +86,6 @@ class TestPluginsCommand:
                 ],
                 "fastapi_root_middlewares": [
                     {
-                        "args": [],
-                        "kwargs": {},
                         "middleware": "unit.plugins.test_plugin.DummyMiddleware",
                         "name": "Name of the Middleware",
                     }
